[kcms/lnf] Make a list variable const
The dataPath variable is used in an extended for-loop as of C++11.
As the variable was not const the for loop uses the more expensive
iterators which allow modification of the list. By making the list
const it can use the better const iterator.

See http://blog.qt.digia.com/blog/2011/05/26/cpp0x-in-qt/

REVIEW: 120913