[Look and feel KCM] Keep cursor size when applying theme
I have configured my system to use an 48 px mouse cursor. When applying a new look and feel theme, it might change the cursor
at runtime but load it with default size. Since the size is never actually changed, on next startup the previous size is restored,
but it should apply it correctly already.

Differential Revision: https://phabricator.kde.org/D10055